chore: setup and run eslint and prettier

+++ src/controllers/seller.ts
@@ -1,4 +1,4 @@
-import { Request, Response, NextFunction } from 'express'
+import {Request, Response} from 'express'
 import fs from 'fs'
 import path from 'path'
 import bcrypt from 'bcrypt'
@@ -6,30 +6,29 @@ import bcrypt from 'bcrypt'
 import Seller from '../models/Seller'
 import Company from '../models/Company'
 import formatImage from '../utils/formatImage'
-import { getDate } from '../utils/getDate'
-import { handleObjectId } from '../utils/handleObjectId'
+import {getDate} from '../utils/getDate'
+import {handleObjectId} from '../utils/handleObjectId'
 
-interface List
-{
+interface List {
 	id: string
 	imagem: string
 	nome: string
 	funcao: string | undefined
 }
 
-const sellerController =
-{
-	create: async (req: Request, res: Response) =>
-	{
+const sellerController = {
+	create: async (req: Request, res: Response) => {
 		const image = req.file
-		const {_id, nome, telefones, email, senha, funcao, admin, representadas} = req.body
+		const {_id, nome, telefones, email, senha, funcao, admin, representadas} =
+			req.body
 
 		const password = bcrypt.hashSync(senha, 10)
 		if (!password)
-			return res.status(500).json({message: 'Algo de errado aconteceu durante a encriptação da senha!'})
-		
-		await Seller.create(
-		{
+			return res.status(500).json({
+				message: 'Algo de errado aconteceu durante a encriptação da senha!'
+			})
+
+		await Seller.create({
 			_id: handleObjectId(_id),
 			nome,
 			imagem: image && image.filename,
@@ -44,22 +43,19 @@ const sellerController =
 		return res.status(201).send()
 	},
 
-	update: async (req: Request, res: Response) =>
-	{
+	update: async (req: Request, res: Response) => {
 		const {id} = req.params
 		const image = req.file
 		const {nome, telefones, email, funcao, admin, representadas} = req.body
 
-		interface Update
-		{
+		interface Update {
 			[letter: string]: any
 		}
-		let seller: Update = {}
+		const seller: Update = {}
 
 		seller['_id'] = id
 		if (nome) seller['nome'] = nome
-		if (image)
-		{
+		if (image) {
 			seller['imagem'] = image.filename
 			const previous = await Seller.findById(id)
 			if (previous?.imagem)
@@ -82,28 +78,23 @@ const sellerController =
 		return tmp
 	},
 
-	remove: async (req: Request, res: Response) =>
-	{
+	remove: async (req: Request, res: Response) => {
 		const {id} = req.params
 
 		const company = await Company.findById(id)
-		if (company?.imagem)
-			fs.unlinkSync(path.resolve('uploads', company.imagem))
+		if (company?.imagem) fs.unlinkSync(path.resolve('uploads', company.imagem))
 
 		const tmp = await Seller.findByIdAndDelete(id)
 		res.status(200).send()
 		return tmp
 	},
 
-	list: async (req: Request, res: Response) =>
-	{
-		let list: List[] = []
+	list: async (req: Request, res: Response) => {
+		const list: List[] = []
 		const sellers = await Seller.find()
 
-		const promises = sellers.map(seller =>
-		{
-			list.push(
-			{
+		const promises = sellers.map(seller => {
+			list.push({
 				id: seller._id,
 				imagem: formatImage(seller.imagem),
 				nome: seller.nome,
@@ -115,25 +106,20 @@ const sellerController =
 		return res.json(list)
 	},
 
-	show: async (req: Request, res: Response) =>
-	{
+	show: async (req: Request, res: Response) => {
 		const seller = await Seller.findById(req.params.id)
-		if(seller !== null)
-		{
-			let companies: {id: string, nome_fantasia: string}[] = []
-			const promises = seller.representadas.map(async company =>
-			{
+		if (seller !== null) {
+			const companies: {id: string; nome_fantasia: string}[] = []
+			const promises = seller.representadas.map(async company => {
 				const tmpCompany = await Company.findById(company.id)
-				companies.push(
-				{
+				companies.push({
 					id: company.id,
 					nome_fantasia: String(tmpCompany?.nome_fantasia)
 				})
 			})
 			await Promise.all(promises)
 
-			return res.json(
-			{
+			return res.json({
 				id: seller._id,
 				imagem: formatImage(seller.imagem),
 				nome: seller.nome,
@@ -144,25 +130,23 @@ const sellerController =
 		}
 	},
 
-	raw: async (req: Request, res: Response) =>
-	{
+	raw: async (req: Request, res: Response) => {
 		const sellers = await Seller.find()
 
-		return res.json(sellers.map(seller =>
-		{
-			let tmp = seller
-			tmp.imagem = formatImage(tmp.imagem)
-			return tmp
-		}))
+		return res.json(
+			sellers.map(seller => {
+				const tmp = seller
+				tmp.imagem = formatImage(tmp.imagem)
+				return tmp
+			})
+		)
 	},
-	
-	rawOne: async (req: Request, res: Response) =>
-	{
+
+	rawOne: async (req: Request, res: Response) => {
 		const {id} = req.params
-		
-		let seller = await Seller.findById(id)
-		if (!seller)
-			return res.status(404).json({message: 'seller not found!'})
+
+		const seller = await Seller.findById(id)
+		if (!seller) return res.status(404).json({message: 'seller not found!'})
 
 		seller.imagem = formatImage(seller.imagem)
